Refresh shop on race finish

See merge request seng201-2025/team-019!21

+++ src/main/java/seng201/team019/gui/EndScreenController.java
@@ -46,7 +46,7 @@ public class EndScreenController extends ScreenController {
         }
         playerNameLabel.setText(getGameEnvironment().getName());
         seasonLengthLabel.setText(String.valueOf(getGameEnvironment().getSeasonLength()));
-        totalPrizeMoneyLabel.setText(String.format("%.2f", (int) getGameEnvironment().getTotalPrizeMoney()));
+        totalPrizeMoneyLabel.setText(String.format("%.0f", getGameEnvironment().getTotalPrizeMoney()));
     }
 
     public void onQuitButtonClicked() {
